feat: log request body content for all methods
Per example solution. Exercise 3.8 asked that data be logged for POST
requests (only), but extending this to all methods feels more
consistent.

The practical effect of this change is that the content of both entry
creation and update (i.e., POST and PATCH) requests are logged; GET and
DELETE requests have no content.